+++ R/errors.R
@@ -72,6 +72,14 @@
 # * Provide print_this() and print_parents() to make it easier to define
 #   custom print methods.
 # * Fix annotating our throw() methods with the incorrect `base::`.
+#
+# ### 1.2.1 -- 2020-01-30
+#
+# * Update wording of error printout to be less intimidating, avoid jargon
+# * Use default printing in interactive mode, so RStudio can detect the
+#   error and highlight it.
+# * Add the rethrow_call_with_cleanup function, to work with embedded
+#   cleancall.
 
 err <- local({
 
@@ -193,9 +201,44 @@ err <- local({
       th(cond)
 
     } else {
-      # We print the error message and possibly the stack ourselves,
-      # because the printing in stop() might truncate it.
-      print_on_error(cond)
+
+      if (is_interactive()) {
+        # In interactive mode, we print the error message through
+        # conditionMessage() and also add a note about .Last.error.trace.
+        # R will potentially truncate the error message, so we make sure
+        # that the note is shown. Ideally we would print the error
+        # ourselves, but then RStudio would not highlight it.
+        max_msg_len <- as.integer(getOption("warning.length"))
+        if (is.na(max_msg_len)) max_msg_len <- 1000
+        msg <- conditionMessage(cond)
+        adv <- style_advice(
+          "\nType .Last.error.trace to see where the error occured"
+        )
+        dots <- "\033[0m\n[...]"
+        if (bytes(msg) + bytes(adv) + bytes(dots) + 5L> max_msg_len) {
+          msg <- paste0(
+            substr(msg, 1, max_msg_len - bytes(dots) - bytes(adv) - 5L),
+            dots
+          )
+        }
+        cond$message <- paste0(msg, adv)
+
+      } else {
+        # In non-interactive mode, we print the error + the traceback
+        # manually, to make sure that it won't be truncated by R's error
+        # message length limit.
+        cat("\n", file = stderr())
+        cat(style_error(gettext("Error: ")), file = stderr())
+        out <- capture_output(print(cond))
+        cat(out, file = stderr(), sep = "\n")
+        out <- capture_output(print(cond$trace))
+        cat(out, file = stderr(), sep = "\n")
+
+        # Turn off the regular error printing to avoid printing
+        # the error twice.
+        opts <- options(show.error.messages = FALSE)
+        on.exit(options(opts), add = TRUE)
+      }
 
       # Dropping the classes and adding "duplicate_condition" is a workaround
       # for the case when we have non-exiting handlers on throw()-n
@@ -206,10 +249,6 @@ err <- local({
       # recognize the duplicates from the "duplicate_condition" extra class.
       class(cond) <- c("duplicate_condition", "condition")
 
-      # And then we turn of the regular error printing to avoid printing
-      # the error twice.
-      opts <- options(show.error.messages = FALSE)
-      on.exit(options(opts), add = TRUE)
       stop(cond)
     }
   }
@@ -314,6 +353,37 @@ err <- local({
     withCallingHandlers(
       # do.call to work around an R CMD check issue
       do.call(".Call", list(.NAME, ...)),
+      error = function(e) {
+        e$`_nframe` <- nframe
+        e$call <- call
+        if (inherits(e, "simpleError")) {
+          class(e) <- c("c_error", "rlib_error", "error", "condition")
+        }
+        e$`_ignore` <- list(c(nframe + 1L, sys.nframe() + 1L))
+        throw(e)
+      }
+    )
+  }
+
+  package_env <- topenv()
+
+  #' Version of rethrow_call that supports cleancall
+  #'
+  #' This function is the same as [rethrow_call()], except that it
+  #' uses cleancall's [.Call()] wrapper, to enable resource cleanup.
+  #' See https://github.com/r-lib/cleancall#readme for more about
+  #' resource cleanup.
+  #'
+  #' @noRd
+  #' @param .NAME Compiled function to call, see [.Call()].
+  #' @param ... Function arguments, see [.Call()].
+  #' @return Result of the call.
+
+  rethrow_call_with_cleanup <- function(.NAME, ...) {
+    call <- sys.call()
+    nframe <- sys.nframe()
+    withCallingHandlers(
+      package_env$call_with_cleanup(.NAME, ...),
       error = function(e) {
         e$`_nframe` <- nframe
         e$call <- call
@@ -510,22 +580,6 @@ err <- local({
     invisible(x)
   }
 
-  print_on_error <- function(cond) {
-    cat("\n", file = stderr())
-    cat(style_error(gettext("Error: ")), file = stderr())
-    out <- capture_output(print(cond))
-    cat(out, file = stderr(), sep = "\n")
-    if (is_interactive()) {
-      cat(
-        style_advice("\nSee `.Last.error.trace` for a stack trace.\n"),
-        file = stderr()
-      )
-    } else {
-      out <- capture_output(print(cond$trace))
-      cat(out, file = stderr(), sep = "\n")
-    }
-  }
-
   capture_output <- function(expr) {
     if (has_crayon()) {
       opts <- options(crayon.enabled = crayon::has_color())
@@ -629,6 +683,10 @@ err <- local({
            USE.NAMES = FALSE)
   }
 
+  bytes <- function(x) {
+    nchar(x, type = "bytes")
+  }
+
   # -- printing, styles -------------------------------------------------
 
   has_crayon <- function() "crayon" %in% loadedNamespaces()
@@ -638,7 +696,7 @@ err <- local({
   }
 
   style_advice <- function(x) {
-    if (has_crayon()) crayon::reset(x) else x
+    if (has_crayon()) crayon::silver(x) else x
   }
 
   style_srcref <- function(x) {
@@ -698,3 +756,4 @@ new_error <- err$new_error
 throw     <- err$throw
 rethrow   <- err$rethrow
 rethrow_call <- err$rethrow_call
+rethrow_call_with_cleanup <- err$.internal$rethrow_call_with_cleanup